thumbnail: fix type of max file size from config

+++ thumbnail/thumbnail.cpp
@@ -452,7 +452,7 @@ QImage ThumbnailProtocol::thumbForDirectory(const QUrl& directory)
         // Directories that the directory preview will be propagated into if there is no direct sub-directories
         const KConfigGroup globalConfig(KSharedConfig::openConfig(), "PreviewSettings");
         m_propagationDirectories = globalConfig.readEntry("PropagationDirectories", QStringList() << "VIDEO_TS").toSet();
-        m_maxFileSize = globalConfig.readEntry("MaximumSize", 5 * 1024 * 1024); // 5 MByte default
+        m_maxFileSize = globalConfig.readEntry("MaximumSize", qulonglong(5 * 1024 * 1024)); // 5 MByte default
     }
 
     const int tiles = 2; //Count of items shown on each dimension
